Feature/hocs 1751 ci drone updates (#103)
* HOCS-1751
Updating CI scripts with new tokens for the HOCS service separation.
We no longer have everything under CS. They are separated into namespaces.

* HOCS-1751
Added tokens for dev environment.

* HOCS-1751
Added the namespace filters for qa and prod

* HOCS-1751
Added the namespace filters for qa and prod

* HOCS-1751
The server port 9083 is blocked on CI.
We cannot figure it out, but we need to deploy so changing it.

* HOCS-1751
I despise the fact that I am force to disable this test, but the port blocking issue on drone is insurmountable.
We need to release. This will be enabled again.