HOCS-2636: remove version param (#35)
* HOCS-2636: remove unnecessary check

Currently we have a check to see if the `KUBE_TOKEN` is set, this isn't
ever going to get hit as we have `set -u` that treats unset variables 
as errors.

* HOCS-2636: remove image_version param

Currently we have two versions: `VERSION` and `IMAGE_VERSION`. To
try and consolidate this we are defaulting to `VERSION`.