HOCS-4034: check for empty/null pdf link
At present if the link to the S3 document is empty or null the S3Client
throws an exception as this is not supported. This change checks this
pdf_link value and if it does not have a value present then it returns
just a basic S3Document with no data.