HOCS-4034: key cannot be empty exception (#128)
* HOCS-4034: change consumes and produces

Previously, we were using `APPLICATION_JSON_UTF8_VALUE` for the
'produces' and 'cosumes'. This is now deprecated in favour of
`APPLICATION_JSON_VALUE` as major browsers interpret UTF-8 special
characters.

* HOCS-4034: check for empty/null pdf link

At present if the link to the S3 document is empty or null the S3Client
throws an exception as this is not supported. This change checks this
pdf_link value and if it does not have a value present then it returns
just a basic S3Document with no data.

* HOCS-4034: use new functionality for pure files

This change uses the previous change to protect PDF file requests for
general files that do not have a file_link in the database.